// src/app/shared/models/sls_intruksi.model.ts
export class SlsIntruksi {
	id?: number;
  sales_lokasi_id: number;
  kepada_lokasi_id: number;
  spk_id: number;
  customer_id: number;
  no_transaksi: string;
  tanggal: string;
  alamat_pengiriman: string;
  keterangan: string;
  pic: string;
  periode_kirim_awal: string;
  periode_kirim_akhir: string;
  // produk_id: number;
  jumlah: number;


}